hpcviewer: apply lazy content provider
We had virtual tree/table, but we didn't set lazy content provider.
This commit sets all trees/tables with lazy content provider. However,
this content provider can theoretically save memory, but it cannot
speedup if we use set the font, foreground and background of each row.

VisualVM reports that 50% of total time in sorting a column is to
handle fonts and colors (tested on Linux+GTK).